Added all prerequisites for CS courses.
NOTE: Some of the prerequisites were ambiguous, e.g. A or B and C is
required. I labeled these with "ambiguous" : true so that in the future
they can be examined for correctness.

Also included is a partially parsed bulletin xml. This contains all the
relevant course data from the unparsed bulletin xml, but with
significantly less noise.